Added async for loop to copy files

+++ src/cleanup.ts
@@ -3,16 +3,17 @@ import * as core from '@actions/core'
 
 export async function run(): Promise<void> {
   try {
-    removeTarArchive()
+    await removeTarArchive()
   } catch (error) {
     const err = error as Error
     core.info(err.message)
   }
 }
-function removeTarArchive(): void {
+async function removeTarArchive(): Promise<void> {
   const path = '/tmp/archive.tar.gz'
   try {
-    fs.unlinkSync(path)
+    if(fs.existsSync(path))
+      await fs.unlinkSync(path)
     core.info(`Action archive cleanup done!`)
   } catch (err) {
     core.info(`Cleanup job failed to complete with error: ${err}`)
